(LANG-624) SystemUtils.getJavaVersionAsFloat throws StringIndexOutOfBoundsException on Android runtime/Dalvik VM. Oops, fix return type.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/lang/trunk@949223 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/lang3/SystemUtils.java
@@ -1669,7 +1669,7 @@ public class SystemUtils {
      * 
      * @return the version, for example 131 for Java 1.3.1
      */
-    static float toJavaVersionInt(String version) {
+    static int toJavaVersionInt(String version) {
         return toVersionInt(toJavaVersionIntArray(version, JAVA_VERSION_TRIM_SIZE));
     }
